eegtxtimgaes/EEG2it-main/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.4.0
axial_positional_embedding==0.3.12
Braindecode==0.7
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
CoLT5-attention==0.11.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
docstring-inheritance==2.2.2
einops==0.8.1
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
h5py==3.15.1
huggingface-hub==0.35.3
hydra-core==1.3.2
hydra_zen==0.15.0
hyper-connections==0.2.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
lightning==2.5.5
lightning-utilities==0.15.2
linear-attention-transformer==0.19.1
linformer==0.2.3
local-attention==1.11.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mne==1.10.2
mne-bids==0.17.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
numpy==2.2.6
omegaconf==2.3.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
product_key_memory==0.2.11
propcache==0.4.1
protobuf==6.33.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.5
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentry-sdk==2.42.0
six==1.17.0
skorch==1.2.0
smmap==5.0.2
soundfile==0.13.1
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0+cu126
torchinfo==1.8.0
torchmetrics==1.8.2
torchvision==0.24.0+cu126
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.22.2
wfdb==4.3.0
yarl==1.22.0
